whodo-populator/requirements.txt
outscraper==5.1.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
unidecode==1.3.8
phonenumbers==8.13.27
